+++ CHANGELOG.md
@@ -7,6 +7,7 @@
     * Because of salesforce-alm removal, the run DX command menu option has been removed.
     * Switched back to using webpack to pack the extension
     * Added ability to cancel running ForceCode Tasks!
+    * Show error message with a link to download if the SFDX CLI isn't installed.
 * 3.16.0
     * Fix issue #264 (s.saveApex(...).then(...).then(...).catch(...).finally is not a function)
     * Fix deploy errors not being reported on first save after file creation